+++ pm4py/vis.py
@@ -26,6 +26,7 @@ from typing import Union, List, Dict, Any, Tuple, Set
 import pandas as pd
 
 from pm4py.objects.bpmn.obj import BPMN
+from pm4py.objects.powl.obj import POWL
 from pm4py.objects.heuristics_net.obj import HeuristicsNet
 from pm4py.objects.log.obj import EventLog, EventStream
 from pm4py.objects.petri_net.obj import PetriNet, Marking
@@ -1263,6 +1264,63 @@ def save_vis_footprints(footprints: Union[Tuple[Dict[str, Any], Dict[str, Any]],
         gviz = fps_visualizer.apply(footprints[0], footprints[1], variant=fps_visualizer.Variants.COMPARISON_SYMMETRIC, parameters={"format": format})
 
     fps_visualizer.save(gviz, file_path)
+
+
+def view_powl(powl: POWL, format: str = constants.DEFAULT_FORMAT_GVIZ_VIEW, bgcolor: str = "white", rankdir: str = "TB"):
+    """
+    Perform a visualization of a POWL model.
+
+    Reference paper:
+    Kourani, Humam, and Sebastiaan J. van Zelst. "POWL: partially ordered workflow language." International Conference on Business Process Management. Cham: Springer Nature Switzerland, 2023.
+
+    :param powl: POWL model
+    :param format: format of the visualization (default: png)
+    :param bgcolor: background color of the visualization (default: white)
+    :param rankdir: sets the direction of the graph ("LR" for left-to-right; "TB" for top-to-bottom)
+
+     .. code-block:: python3
+
+        import pm4py
+
+        log = pm4py.read_xes('tests/input_data/running-example.xes')
+        powl_model = pm4py.discover_powl(log)
+        pm4py.view_powl(powl_model, format='svg')
+    """
+    format = str(format).lower()
+
+    from pm4py.visualization.powl import visualizer as powl_visualizer
+    gviz = powl_visualizer.apply(powl, parameters={"format": format, "bgcolor": bgcolor, "rankdir": rankdir})
+
+    powl_visualizer.view(gviz)
+
+
+def save_vis_powl(powl: POWL, file_path: str, bgcolor: str = "white", rankdir: str = "TB"):
+    """
+    Saves the visualization of a POWL model.
+
+    Reference paper:
+    Kourani, Humam, and Sebastiaan J. van Zelst. "POWL: partially ordered workflow language." International Conference on Business Process Management. Cham: Springer Nature Switzerland, 2023.
+
+    :param powl: POWL model
+    :param file_path: target path of the visualization
+    :param bgcolor: background color of the visualization (default: white)
+    :param rankdir: sets the direction of the graph ("LR" for left-to-right; "TB" for top-to-bottom)
+
+     .. code-block:: python3
+
+        import pm4py
+
+        log = pm4py.read_xes('tests/input_data/running-example.xes')
+        powl_model = pm4py.discover_powl(log)
+        pm4py.save_vis_powl(powl_model, 'powl.png')
+    """
+    file_path = str(file_path)
+    format = os.path.splitext(file_path)[1][1:].lower()
+
+    from pm4py.visualization.powl import visualizer as powl_visualizer
+    gviz = powl_visualizer.apply(powl, parameters={"format": format, "bgcolor": bgcolor, "rankdir": rankdir})
+
+    powl_visualizer.save(gviz, file_path)
 
 
 def view_object_graph(ocel: OCEL, graph: Set[Tuple[str, str]], format: str = constants.DEFAULT_FORMAT_GVIZ_VIEW, bgcolor: str = "white", rankdir: str = constants.DEFAULT_RANKDIR_GVIZ):